구조 리팩터링
